initial spike of coffeejade support

+++ src/main/java/com/voltvoodoo/brew/compile/RawCopyCompiler.java
@@ -2,11 +2,21 @@ package com.voltvoodoo.brew.compile;
 
 import org.codehaus.plexus.util.IOUtil;
 
-import java.io.*;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
 import java.util.List;
 
 public class RawCopyCompiler implements Compiler
 {
+    private final CompilerMojo mojo;
+
+    public RawCopyCompiler(CompilerMojo mojo)
+    {
+        this.mojo = mojo;
+    }
+
     public void compile(List<String> files, File sourceDir, File targetDir)
     {
         for (String path : files)
@@ -22,21 +32,24 @@ public class RawCopyCompiler implements Compiler
         }
     }
 
-    public static void copyFile(String path, File sourceDir, File targetDir) throws IOException
+    public void copyFile(String path, File sourceDir, File targetDir) throws IOException
     {
         FileOutputStream out = null;
         FileInputStream in = null;
         try
         {
             File source = new File(sourceDir, path);
-            File target = new File(targetDir, path);
-            File parentDir = target.getParentFile();
-            parentDir.mkdirs();
-            if (parentDir.exists())
+            if (source.exists() && !mojo.isSourceFile(source))
             {
-                in = new FileInputStream(source);
-                out = new FileOutputStream(target);
-                IOUtil.copy(in, out);
+                File target = new File(targetDir, path);
+                File parentDir = target.getParentFile();
+                parentDir.mkdirs();
+                if (parentDir.exists())
+                {
+                    in = new FileInputStream(source);
+                    out = new FileOutputStream(target);
+                    IOUtil.copy(in, out);
+                }
             }
         }
         finally
